break out of the right loop
If a complete packet is received while we're sending a multi-part
message, at the end we break out of the wrong loop and send $item --
actually still an array ref -- down the socket, where it gets
stringified into ARRAY(0xdeadbeef).  This is almost certainly not what
the other end is expecting.